Add SignMessage middleware
